# task5/repo-stat/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing repository subscriptions and fetching statistics",
        "title": "Repository Stats API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/ping": {
            "get": {
                "description": "Check the health status of internal services (processor, subscriber)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Ping services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PingResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repositories/info": {
            "get": {
                "description": "Returns statistics for a given GitHub repository URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "Get repository information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GitHub Repository URL",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RepoInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions": {
            "get": {
                "description": "Returns a list of all repositories currently subscribed by the user from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get list of all active subscriptions",
                "responses": {
                    "200": {
                        "description": "List of subscribed repositories",
                        "schema": {
                            "$ref": "#/definitions/dto.ListSubscriptionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a subscription to a GitHub repository. Validates that the repository exists on GitHub before saving to the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Add a new repository subscription",
                "parameters": [
                    {
                        "description": "Owner and Repository name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Subscription added successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.AddSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format or repository not found on GitHub",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/info": {
            "get": {
                "description": "Fetches real-time statistics (stars, forks, description) from GitHub for every repository in the subscription list.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Get statistics for all subscribed repositories",
                "responses": {
                    "200": {
                        "description": "Statistics for all subscribed repos",
                        "schema": {
                            "$ref": "#/definitions/dto.SubscriptionsInfoResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/subscriptions/{owner}/{repo}": {
            "delete": {
                "description": "Removes a subscription for a given owner and repository name from the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Remove a specific subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GitHub Owner username",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription removed successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.RemoveSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.PingStatus": {
            "type": "string",
            "enum": [
                "up",
                "down"
            ],
            "x-enum-varnames": [
                "PingStatusUp",
                "PingStatusDown"
            ]
        },
        "domain.ServicePingResult": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.PingStatus"
                }
            }
        },
        "dto.AddSubscriptionRequest": {
            "type": "object",
            "properties": {
                "owner": {
                    "type": "string"
                },
                "repo": {
                    "type": "string"
                }
            }
        },
        "dto.AddSubscriptionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.ListSubscriptionsResponse": {
            "type": "object",
            "properties": {
                "subscriptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SubscriptionItem"
                    }
                }
            }
        },
        "dto.PingResponse": {
            "type": "object",
            "properties": {
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ServicePingResult"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.RemoveSubscriptionResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        },
        "dto.RepoInfoResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "forks": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string"
                },
                "stars": {
                    "type": "integer"
                }
            }
        },
        "dto.RepoStatItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "forks": {
                    "type": "integer"
                },
                "full_name": {
                    "type": "string"
                },
                "stars": {
                    "type": "integer"
                }
            }
        },
        "dto.SubscriptionItem": {
            "type": "object",
            "properties": {
                "owner": {
                    "type": "string"
                },
                "repo": {
                    "type": "string"
                }
            }
        },
        "dto.SubscriptionsInfoResponse": {
            "type": "object",
            "properties": {
                "stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RepoStatItem"
                    }
                }
            }
        }
    }
}